requirements.txt
torchvision==0.5.0
pytorch_lightning==0.8.5
torch==1.4.0
scikit_learn==0.22.1
Pillow==7.0.0
attrs==19.3.0